shared/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "tickdeck-shared"
version = "0.1.0"
description = "TickDeck 공유 서비스 (crawler, gemini, pptx_builder)"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "python-pptx>=0.6.23",
    "google-genai>=0.5.0",
    "pydantic>=2.7.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["shared*"]
